GW-71: Refactored GridContainer for event
